move NoteId structure
