+++ zcash_client_sqlite/src/lib.rs
@@ -35,7 +35,7 @@
 extern crate core;
 
 use std::collections::HashMap;
-use std::fmt;
+//use std::fmt;
 use std::path::Path;
 
 use rusqlite::{Connection, Statement};
@@ -58,6 +58,7 @@ use zcash_client_backend::{
     proto::compact_formats::CompactBlock,
     wallet::{AccountId, SpendableNote},
 };
+use zcash_extras::NoteId;
 
 use crate::error::SqliteClientError;
 
@@ -66,22 +67,22 @@ pub mod error;
 pub mod for_async;
 pub mod wallet;
 
-/// A newtype wrapper for sqlite primary key values for the notes
-/// table.
-#[derive(Debug, Copy, Clone)]
-pub enum NoteId {
-    SentNoteId(i64),
-    ReceivedNoteId(i64),
-}
-
-impl fmt::Display for NoteId {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            NoteId::SentNoteId(id) => write!(f, "Sent Note {}", id),
-            NoteId::ReceivedNoteId(id) => write!(f, "Received Note {}", id),
-        }
-    }
-}
+///// A newtype wrapper for sqlite primary key values for the notes
+///// table.
+//#[derive(Debug, Copy, Clone)]
+//pub enum NoteId {
+//    SentNoteId(i64),
+//    ReceivedNoteId(i64),
+//}
+//
+//impl fmt::Display for NoteId {
+//    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+//        match self {
+//            NoteId::SentNoteId(id) => write!(f, "Sent Note {}", id),
+//            NoteId::ReceivedNoteId(id) => write!(f, "Received Note {}", id),
+//        }
+//    }
+//}
 
 /// A wrapper for the SQLite connection to the wallet database.
 pub struct WalletDb<P> {
